pyproject.toml
[project]
name = "vhr-geo-fm"
version = "0.0.1"
description = "VHR Foundation Model evaluation via GEO-Bench-2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geobenchv2>=0.9",
    "hydra-core>=1.3",
    "lightning>=2.0",
    "optuna>=4.0",
    "optuna-integration[pytorch-lightning]>=4.8.0",
    "pandas>=2.0",
    "scipy>=1.14",
    "tabulate>=0.9",
    "terratorch>=1.2.4",
    "timm>=1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0",
    "pytest>=8.0",
    "ruff>=0.14",
    "ty>=0.0.26",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ty.src]
exclude = ["scripts/patch_geo_bench.py"]

[tool.uv.build-backend]
module-name = "eval"
module-root = ""

[build-system]
requires = ["uv_build>=0.7"]
build-backend = "uv_build"
